+++ src/App/Main.fs
@@ -22,13 +22,14 @@ type ISavedState =
     abstract code: string
     abstract html: string
     abstract css: string
+    abstract outputCodeActive: bool
 
 let private Worker(): Worker =
     // importDefault "worker-loader!../../../Fable/src/fable-standalone/src/Worker/Worker.fsproj"
     Worker.Create(Literals.WORKER_BUNDLE_URL)
 
 let private loadState(_key: string): ISavedState = importMember "./js/util.js"
-let private saveState(_key: string, _code: string, _html: string, _cssCode : string): unit = importMember "./js/util.js"
+let private saveState(_key: string, _code: string, _html: string, _cssCode : string, _outputCodeActive: bool): unit = importMember "./js/util.js"
 let private updateQuery(_fsharpCode : string, _htmlCode : string, _cssCode : string): unit = importMember "./js/util.js"
 
 type IEditor = Monaco.Editor.IStandaloneCodeEditor
@@ -68,7 +69,7 @@ type Model =
         IFrameUrl : string
         OutputTab : OutputTab
         CodeTab : CodeTab
-        CodeES2015: string
+        CompiledCode: string
         FSharpCode : string
         FSharpErrors : Monaco.Editor.IMarkerData[]
         HtmlCode: string
@@ -80,7 +81,7 @@ type Model =
         Logs : ConsolePanel.Log list
     }
 
-type EndCompileStatus = Result<string * Fable.Standalone.Error[], string>
+type EndCompileStatus = Result<string * string * Fable.Standalone.Error[], string>
 
 type Msg =
     | SetFSharpEditor of IEditor
@@ -146,8 +147,11 @@ let private addLog log (model : Model) =
                 model.Logs @ [log]
     }
 
-let private generateHtmlUrl (model: Model) jsCode =
-    saveState(Literals.STORAGE_KEY, model.FSharpCode, model.HtmlCode, model.CssCode)
+let private saveModel (model: Model) =
+    saveState(Literals.STORAGE_KEY, model.FSharpCode, model.HtmlCode, model.CssCode, model.OutputTab = OutputTab.Code)
+
+let private saveModelAndGenerateHtmlUrl (model: Model) jsCode =
+    saveModel model
     Generator.generateHtmlBlobUrl model.HtmlCode model.CssCode jsCode
 
 let private clamp min max value =
@@ -306,7 +310,7 @@ let update msg (model : Model) =
             FSharpCode = saved.code
             HtmlCode = saved.html
             CssCode = saved.css
-            CodeES2015 = ""
+            CompiledCode = ""
             IFrameUrl = ""
             Logs = []
         }
@@ -370,7 +374,8 @@ let update msg (model : Model) =
                 | Some code -> code
                 | None -> model.FSharpCode
             let opts = model.Sidebar.Options
-            CompileCode(code, opts.ToOtherFSharpOptions) |> model.Worker.Post
+            let language = model.Sidebar.Options.Language
+            CompileCode(code, language, opts.ToOtherFSharpOptions) |> model.Worker.Post
 
             { model with
                 State = Compiling
@@ -379,7 +384,7 @@ let update msg (model : Model) =
 
     | EndCompile result ->
         match result with
-        | Ok(codeES2015, errors) ->
+        | Ok(compiledCode, lang, errors) ->
             let hasCriticalErrors = errors |> Array.exists (fun e -> not e.IsWarning)
             if hasCriticalErrors then
                 let toastCmd =
@@ -397,7 +402,12 @@ let update msg (model : Model) =
                 , toastCmd
 
             else
-                let toastCmd =
+                let cmd1 =
+                    match lang.ToLower() with
+                    | "js" | "javascript" -> Cmd.OfFunc.perform (saveModelAndGenerateHtmlUrl model) compiledCode SetIFrameUrl
+                    | _ -> [fun _ -> saveModel model]
+
+                let cmd2 =
                     Toast.message "Compiled successfuly"
                     |> Toast.position Toast.BottomRight
                     |> Toast.icon Fa.Solid.Check
@@ -405,15 +415,12 @@ let update msg (model : Model) =
                     |> Toast.success
 
                 { model with
-                    CodeES2015 = codeES2015
+                    CompiledCode = compiledCode
                     State = Compiled
                     FSharpErrors = mapErrorToMarker errors
                     Logs = [ ConsolePanel.Log.Separator ]
                 }
-                , Cmd.batch [
-                    Cmd.OfFunc.perform (generateHtmlUrl model) codeES2015 SetIFrameUrl
-                    toastCmd
-                ]
+                , Cmd.batch [ cmd1; cmd2 ]
 
         | Error msg ->
             { model with
@@ -586,7 +593,7 @@ let update msg (model : Model) =
     | RefreshIframe ->
         let jsCode = model.JsEditor.getValue()
         { model with
-            CodeES2015 = jsCode
+            CompiledCode = jsCode
             Logs = [ ConsolePanel.Log.Separator ]
         }
         , Cmd.OfFunc.perform (Generator.generateHtmlBlobUrl model.HtmlCode model.CssCode) jsCode SetIFrameUrl
@@ -614,11 +621,11 @@ let workerCmd (worker : ObservableWorker<_>)=
                 LoadSuccess version |> dispatch
             | LoadFailed -> LoadFail |> dispatch
             | ParsedCode errors -> MarkEditorErrors errors |> dispatch
-            | CompilationFinished (jsCode, errors, stats) ->
-                Ok(jsCode, errors) |> EndCompile |> dispatch
+            | CompilationFinished (code, lang, errors, stats) ->
+                Ok(code, lang, errors) |> EndCompile |> dispatch
                 UpdateStats stats |> dispatch
-            | CompilationsFinished (jsCode, errors, stats) ->
-                Ok(Array.last jsCode, errors) |> EndCompile |> dispatch
+            | CompilationsFinished (code, lang, errors, stats) ->
+                Ok(Array.last code, lang, errors) |> EndCompile |> dispatch
                 UpdateStats stats |> dispatch
             | CompilerCrashed msg -> Error msg |> EndCompile |> dispatch
             // Do nothing, these will be handled by .PostAndAwaitResponse
@@ -652,9 +659,9 @@ let init () =
         JsEditor = Unchecked.defaultof<IEditor>
         Worker = worker
         IFrameUrl = ""
-        OutputTab = OutputTab.Live
+        OutputTab = if saved.outputCodeActive then OutputTab.Code else OutputTab.Live
         CodeTab = CodeTab.FSharp
-        CodeES2015 = ""
+        CompiledCode = ""
         FSharpCode = saved.code
         FSharpErrors = [||]
         HtmlCode = saved.html
@@ -1034,7 +1041,7 @@ let private viewCodeEditor (model: Model) dispatch =
                         let minimapOptions = jsOptions<Monaco.Editor.IEditorMinimapOptions>(fun oMinimap ->
                             oMinimap.enabled <- Some false
                         )
-                        o.language <- Some "javascript"
+                        // o.language <- Some "javascript"
                         o.fontSize <- Some model.Sidebar.Options.FontSize
                         o.theme <- Some "vs-dark"
                         o.minimap <- Some minimapOptions
@@ -1045,7 +1052,7 @@ let private viewCodeEditor (model: Model) dispatch =
 
     ReactEditor.editor [
         editor.options options
-        editor.value model.CodeES2015
+        editor.value model.CompiledCode
         editor.isHidden (model.OutputTab <> OutputTab.Code)
         editor.customClass (fontSizeClass model.Sidebar.Options.FontSize)
         editor.editorDidMount (onJsEditorDidMount model dispatch)