+++ src/App/Widgets/Options.fs
@@ -18,6 +18,7 @@ type Model =
     { Optimize : bool
       DefineDebug : bool
       TypedArrays : bool
+      Language : string
       FontSize : float
       FontFamily : string
       GistToken : string option
@@ -32,6 +33,7 @@ type Model =
         { Optimize = false
           DefineDebug = true
           TypedArrays = true
+          Language = "javascript"
           FontSize = 14.
           FontFamily = MONACO_DEFAULT_FONT_FAMILY
           GistToken = None
@@ -45,6 +47,7 @@ type Model =
                         //     |> Option.defaultValue false
               DefineDebug = get.Optional.Field "defineDebug" Decode.bool |> Option.defaultValue true
               TypedArrays = get.Optional.Field "typedArrays" Decode.bool |> Option.defaultValue true
+              Language = get.Optional.Field "language" Decode.string |> Option.defaultValue "javascript"
               FontSize = get.Optional.Field "fontSize" Decode.float
                             |> Option.defaultValue 14.
               FontFamily = get.Optional.Field "fontFamily" Decode.string
@@ -58,6 +61,7 @@ type Model =
             [ yield "optimize", Encode.bool model.Optimize
               yield "defineDebug", Encode.bool model.DefineDebug
               yield "typedArrays", Encode.bool model.TypedArrays
+              yield "language", Encode.string model.Language
               yield "fontSize", Encode.float model.FontSize
               yield "fontFamily", Encode.string model.FontFamily
               match model.GistToken with
@@ -67,13 +71,18 @@ type Model =
 type Msg =
     | ToggleOptimize
     | ToggleDefineDebug
-    | TogglePreviewLanguage
+    | ToggleTypedArrays
+    | ChangeLanguage of string
     | ChangeFontSize of float
     | ChangeFontFamily of string
     | ChangeGistToken of string
     | SaveToken
     | DeleteToken
 
+[<RequireQualifiedAccess>]
+type ExtMsg =
+    | NoOp
+    | Recompile
 
 let init () =
     match localStorage.getItem(LOCAL_STORAGE_REPL_SETTING) with
@@ -88,40 +97,43 @@ let init () =
             JS.console.log("Error while loading your settings from localStorage:\n", msg)
             Model.Default
 
-let saveSettings (model : Model) =
+let saveSettings (model : Model, extMsg: ExtMsg) =
     let data =
         Model.Encoder model
         |> Encode.toString 0
 
     localStorage.setItem(LOCAL_STORAGE_REPL_SETTING, data)
-    model
+    model, extMsg
 
 let update msg model =
     match msg with
     | ToggleOptimize ->
-        { model with Optimize = not model.Optimize }
+        { model with Optimize = not model.Optimize }, ExtMsg.NoOp
 
     | ToggleDefineDebug ->
-        { model with DefineDebug = not model.DefineDebug }
+        { model with DefineDebug = not model.DefineDebug }, ExtMsg.NoOp
 
-    | TogglePreviewLanguage ->
-        { model with TypedArrays = not model.TypedArrays }
+    | ToggleTypedArrays ->
+        { model with TypedArrays = not model.TypedArrays }, ExtMsg.NoOp
+
+    | ChangeLanguage newLang ->
+        { model with Language = newLang }, ExtMsg.Recompile
 
     | ChangeFontSize newSize ->
-        { model with FontSize = newSize }
+        { model with FontSize = newSize }, ExtMsg.NoOp
 
     | ChangeFontFamily newFont ->
-        { model with FontFamily = newFont }
+        { model with FontFamily = newFont }, ExtMsg.NoOp
 
     | ChangeGistToken token ->
-        { model with GistTokenField = token }
+        { model with GistTokenField = token }, ExtMsg.NoOp
 
     | SaveToken ->
         { model with GistTokenField = ""
-                     GistToken = Some model.GistTokenField }
+                     GistToken = Some model.GistTokenField }, ExtMsg.NoOp
 
     | DeleteToken ->
-        { model with GistToken = None}
+        { model with GistToken = None}, ExtMsg.NoOp
 
     // Save the setting in localStorage
     // Like that they are persistant between REPL reload
@@ -178,6 +190,33 @@ let inline private fontFamilySetting (fontFamily : string) dispatch =
         ]
     ]
 
+let private languageSetting (language : string) dispatch =
+    let option (txt: string) =
+        Html.option [
+            prop.value txt
+            prop.text txt
+        ]
+
+    Bulma.field.div [
+        Bulma.label "Language"
+
+        Bulma.control.div [
+            Bulma.select [
+                select.isFullWidth
+                prop.value language
+                prop.onChange (fun (ev : Types.Event) ->
+                    ev.Value |> ChangeLanguage |> dispatch
+                )
+                prop.children [
+                    option "JavaScript"
+                    option "TypeScript"
+                    option "Python"
+                    option "Rust"
+                    option "Dart"
+                ]
+            ]
+        ]
+    ]
 
 let private switchOption (label : string) isActive dispatch msg =
     Bulma.field.div [
@@ -206,8 +245,8 @@ let inline private optimizeSetting (model: Model) dispatch =
 let private defineDebugSetting (model: Model) dispatch =
     switchOption "Define DEBUG" model.DefineDebug dispatch ToggleDefineDebug
 
-let private previewLanguageSetting (model: Model) dispatch =
-    switchOption "Typed Arrays" model.TypedArrays dispatch TogglePreviewLanguage
+let private typedArraysSetting (model: Model) dispatch =
+    switchOption "Typed Arrays" model.TypedArrays dispatch ToggleTypedArrays
 
 let inline private gistTokenSetting (token : string option) (tokenField : string) dispatch =
     match token with
@@ -254,11 +293,15 @@ let inline private gistTokenSetting (token : string option) (tokenField : string
 
 let view (model: Model) dispatch =
     Html.div [
-        defineDebugSetting model dispatch
-        previewLanguageSetting model dispatch
-        fontFamilySetting model.FontFamily dispatch
-        fontSizeSetting model.FontSize dispatch
-        gistTokenSetting model.GistToken model.GistTokenField dispatch
-        // TODO: Optimize is disabled to prevent problems with inline functions in REPL Lib
-        //   optimizeSetting model dispatch
+        prop.className "options"
+        prop.children [
+            defineDebugSetting model dispatch
+            typedArraysSetting model dispatch
+            languageSetting model.Language dispatch
+            fontFamilySetting model.FontFamily dispatch
+            fontSizeSetting model.FontSize dispatch
+            gistTokenSetting model.GistToken model.GistTokenField dispatch
+            // TODO: Optimize is disabled to prevent problems with inline functions in REPL Lib
+            //   optimizeSetting model dispatch
+        ]
     ]